+++ ts_runtime/src/peer_tracker/peer_db.rs
@@ -545,6 +545,7 @@ mod test {
             peerapi_dns_proxy: false,
             is_wireguard_only: false,
             exit_node_dns_resolvers: vec![],
+            service_vips: Default::default(),
         }
     }
 
@@ -830,6 +831,7 @@ mod test {
                     peerapi_dns_proxy: false,
                     is_wireguard_only: false,
                     exit_node_dns_resolvers: vec![],
+                    service_vips: Default::default(),
                 }
             })
             .collect()